Indexing Fix
There was a bug in streams.binaryStreamGen where an input hologram array which wasn't square would throw index out of bound errors. I think I fixed it, but my method was based on heuristically changing things until they worked. If this breaks in the future this is a good point to go back to and see what changed.